requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docutils==0.21.2
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
markdown2==2.5.3
MarkupSafe==3.0.2
multidict==6.1.0
nicegui==2.11.0
numpy==2.2.2
opencv-python==4.11.0.86
orjson==3.10.15
propcache==0.2.1
pscript==0.7.7
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.0.4
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
